+++ src/product/services.js
@@ -1,4 +1,6 @@
 const Product = require('../models/product_model');
+const Admin = require('../models/admin_model');
+const createError = require('http-errors');
 const Joi = require('joi');
 
 const schema = Joi.object({
@@ -20,17 +22,29 @@ const schema = Joi.object({
                 
 });
 
-const getAll = async () => {
-    let products = await Product.find();
+const getAll = async (admin) => {
+    let products = await Product.find().where("admin").equals(admin);
     return products;
 };
 
-const getById = async (id) => {
-    let product = await Product.find( { "_id": id} );
+const getAllUser = async (name) => {
+    let admin = await Admin.findOne( { name } );
+    let products = await Product.find({ "admin" : admin._id });
+    return products;
+};
+
+const getById = async (id, admin_id) => {
+    let product = await Product.findById(id);
+    if (!product){
+        throw new createError(400, `Not found`);
+    }
+    if( product.admin._id.toString() !== admin_id){
+        throw new createError(400, `Invalid action`);
+    }
     return product;
 }
 
-const create = async (body) => {
+const create = async (id, body) => {
     let product = new Product({
         name: body.name,
         price: body.price,
@@ -38,29 +52,42 @@ const create = async (body) => {
         stock: body.stock,
         image: body.image,
     });
+    product.admin = id;
     return await product.save();
 }
 
-const update = async (id, {name, price, description, stock, image} ) => {
-    let product = await Product.findOneAndUpdate( { "_id" : id }, {
-        $set: {
-            "name": name,
-            "price": price,
-            "description": description,
-            "stock": stock,
-            "image": image
-        }
-    }, {new: true});
-    return product;
+const update = async (id, {name, price, description, stock, image}, admin_id ) => {
+    const product = await Product.findById( id );
+    if (!product){
+        throw new createError(400, `Not found`);
+    }
+    if( product.admin._id.toString() !== admin_id){
+        throw new createError(400, `Invalid action`);
+    }
+    product.name = name || product.name;
+    product.price = price || product.price;
+    product.description = description || product.description;
+    product.stock = stock || product.stock;
+    product.image = image || product.image;
+    const newProduct = await product.save();
+    return newProduct;
 }
 
-const deleteP = async (id) => {
-    let product = await Product.findOneAndDelete( { "_id" : id});
-    return product;
+const deleteP = async (id, admin_id) => {
+    const product = await Product.findById( id );
+    if (!product){
+        throw new createError(400, `Not found`);
+    }
+    if( product.admin._id.toString() !== admin_id){
+        throw new createError(400, `Invalid action`);
+    }
+    const newProduct = await Product.findOneAndDelete( { "_id" : id});
+    return newProduct;
 };
 
 module.exports.ProductService = {
     getAll,
+    getAllUser,
     getById,
     create,
     update,
